+++ gala/potential/potential/builtin/builtin_potentials.h
@@ -92,3 +92,7 @@ extern double longmuralibar_value(double t, double *pars, double *q, int n_dim);
 extern void longmuralibar_gradient(double t, double *pars, double *q, int n_dim, double *grad);
 extern double longmuralibar_density(double t, double *pars, double *q, int n_dim);
 extern void longmuralibar_hessian(double t, double *pars, double *q, int n_dim, double *hess);
+
+extern double burkert_value(double t, double *pars, double *q, int n_dim);
+extern void burkert_gradient(double t, double *pars, double *q, int n_dim, double *grad);
+extern double burkert_density(double t, double *pars, double *q, int n_dim);