+++ CHANGES.rst
@@ -7,6 +7,8 @@ New Features
 - Added an option to specify a multiprocessing or parallel processing pool when
   computing basis function coefficients for the SCF potential from discrete particles.
 
+- Added the Burkert potential as a built-in cpotential.
+
 Bug fixes
 ---------